+++ lib/jmap/mail/email/individual_header_identifier.dart
@@ -14,7 +14,6 @@ class IndividualHeaderIdentifier with EquatableMixin {
   static final importanceHeader = IndividualHeaderIdentifier('header:Importance:asText');
   static final listPostHeader = IndividualHeaderIdentifier('header:List-Post:asText');
   static final listUnsubscribeHeader = IndividualHeaderIdentifier('header:List-Unsubscribe:asText');
-  static final xLinkedFileHeader = IndividualHeaderIdentifier('header:X-LinkedFile:all:asText');
 
   final String value;